+++ src/GitHub/api.ts
@@ -325,3 +325,66 @@ export async function deleteGitHubFile(repo: TRepo, file: TContent) {
         output?.logError(repo, e);
     }
 }
+
+/**
+ * Create a new GitHub repository
+ *
+ * @export
+ * @async
+ * @param {string} owner The owner of the repository
+ * @param {string} repoName The name of the repository
+ * @param {boolean} isPrivate Whether the repository should be private
+ * @returns {(Promise<TRepo | undefined>)}
+ */
+export async function newGitHubRepository(owner: string, repoName: string, isPrivate: boolean): Promise<TRepo | undefined> {
+    const octokit = new rest.Octokit({
+        auth: await credentials.getAccessToken(),
+    });
+
+    try {
+        let newRepo = await octokit.repos.createForAuthenticatedUser({
+            name: repoName,
+            private: isPrivate,
+            auto_init: true,
+            headers: {
+                Accept: "application/vnd.github.v3+json",
+            },
+        });
+
+        return Promise.resolve(newRepo.data);
+    } catch (e: any) {
+        output?.appendLine(`${e.message}: ${owner}/${repoName}`, output.messageType.error);
+    }
+
+    return Promise.reject(undefined);
+}
+
+/**
+ * Delete a GitHub repository
+ * @date 9/26/2022 - 9:39:39 AM
+ *
+ * @export
+ * @async
+ * @param {TRepo} repo The repository to delete
+ * @returns {Promise<boolean>}
+ */
+// export async function deleteGitHubRepository(repo: TRepo): Promise<boolean>;
+// export async function deleteGitHubRepository(owner: string, repoName: string): Promise<boolean>;
+export async function deleteGitHubRepository(repo: TRepo): Promise<boolean> {
+    const octokit = new rest.Octokit({
+        auth: await credentials.getAccessToken(),
+    });
+
+    try {
+        await octokit.repos.delete({
+            owner: repo.owner.login,
+            repo: repo.name,
+        });
+
+        return Promise.resolve(true);
+    } catch (e: any) {
+        output?.logError(repo, e);
+    }
+
+    return Promise.reject(false);
+}