+++ src/extension.ts
@@ -4,7 +4,7 @@ import * as trace from "./tracing";
 import { commands, ExtensionContext, workspace, window, ProgressLocation } from "vscode";
 import { ContentNode, RepoProvider } from "./Tree/nodes";
 import { RepoFileSystemProvider, REPO_SCHEME } from "./FileSystem/fileSystem";
-import { addFile, deleteNode, pickRepository, uploadFiles } from "./GitHub/commands";
+import { addFile, deleteNode, deleteRepository, newRepository, pickRepository, uploadFiles } from "./GitHub/commands";
 import { TGitHubUser } from "./GitHub/types";
 import { addToGlobalStorage, clearGlobalStorage, getReposFromGlobalStorage, purgeGlobalStorage, removeFromGlobalStorage, store } from "./FileSystem/storage";
 import { GLOBAL_STORAGE_KEY } from "./GitHub/constants";
@@ -85,6 +85,18 @@ export async function activate(context: ExtensionContext) {
     context.subscriptions.push(
         commands.registerCommand("VirtualRepos.downloadRepository", async (node) => {
             throw new Error("Not implemented");
+        })
+    );
+
+    context.subscriptions.push(
+        commands.registerCommand("VirtualRepos.newRepository", async () => {
+            newRepository(true);
+        })
+    );
+
+    context.subscriptions.push(
+        commands.registerCommand("VirtualRepos.deleteRepository", async (node) => {
+            deleteRepository(node);
         })
     );